Use long for Buffer.Length
This would be necessary for databases bigger than 2 GB. I suspect the
memory-mapped mode would mostly work for such databases.

When using memory-mode, the decoding of strings will likely be broken
due to the use of an Int32 for Encoding.UTF8.GetString. There is no
Int64 equivalent for the API. A possible workaround would be to pass it
a Span<byte> instead. I don't think it is worth pursuing at this time
though as all available databases are much smaller.